+++ packages/client/src/Call.ts
@@ -90,10 +90,7 @@ import {
   VideoTrackType,
 } from './types';
 import { BehaviorSubject, Subject, takeWhile } from 'rxjs';
-import {
-  ReconnectDetails,
-  VideoLayerSetting,
-} from './gen/video/sfu/event/events';
+import { ReconnectDetails } from './gen/video/sfu/event/events';
 import {
   ClientDetails,
   TrackType,
@@ -205,6 +202,7 @@ export class Call {
    */
   private readonly dispatcher = new Dispatcher();
 
+  private publishOptions?: PublishOptions;
   private statsReporter?: StatsReporter;
   private sfuStatsReporter?: SfuStatsReporter;
   private dropTimeout: ReturnType<typeof setTimeout> | undefined;
@@ -521,7 +519,7 @@ export class Call {
         await waitUntilCallJoined();
       }
 
-      if (this.ringing) {
+      if (reject && this.ringing) {
         // I'm the one who started the call, so I should cancel it.
         const hasOtherParticipants = this.state.remoteParticipants.length > 0;
         if (
@@ -532,7 +530,7 @@ export class Call {
           // Signals other users that I have cancelled my call to them
           // before they accepted it.
           await this.reject();
-        } else if (reject && callingState === CallingState.RINGING) {
+        } else if (callingState === CallingState.RINGING) {
           // Signals other users that I have rejected the incoming call.
           await this.reject();
         }
@@ -1296,19 +1294,12 @@ export class Call {
           break;
         case TrackType.VIDEO:
           const videoStream = this.camera.state.mediaStream;
-          if (videoStream) {
-            await this.publishVideoStream(
-              videoStream,
-              this.camera.publishOptions,
-            );
-          }
+          if (videoStream) await this.publishVideoStream(videoStream);
           break;
         case TrackType.SCREEN_SHARE:
           const screenShareStream = this.screenShare.state.mediaStream;
           if (screenShareStream) {
-            await this.publishScreenShareStream(screenShareStream, {
-              screenShareSettings: this.screenShare.getSettings(),
-            });
+            await this.publishScreenShareStream(screenShareStream);
           }
           break;
         // screen share audio can't exist without a screen share, so we handle it there
@@ -1340,12 +1331,8 @@ export class Call {
    * The previous video stream will be stopped.
    *
    * @param videoStream the video stream to publish.
-   * @param opts the options to use when publishing the stream.
    */
-  publishVideoStream = async (
-    videoStream: MediaStream,
-    opts: PublishOptions = {},
-  ) => {
+  publishVideoStream = async (videoStream: MediaStream) => {
     if (!this.sfuClient) throw new Error(`Call not joined yet.`);
     // joining is in progress, and we should wait until the client is ready
     await this.sfuClient.joinTask;
@@ -1367,7 +1354,7 @@ export class Call {
       videoStream,
       videoTrack,
       TrackType.VIDEO,
-      opts,
+      this.publishOptions,
     );
   };
 
@@ -1411,12 +1398,8 @@ export class Call {
    * The previous screen-share stream will be stopped.
    *
    * @param screenShareStream the screen-share stream to publish.
-   * @param opts the options to use when publishing the stream.
    */
-  publishScreenShareStream = async (
-    screenShareStream: MediaStream,
-    opts: PublishOptions = {},
-  ) => {
+  publishScreenShareStream = async (screenShareStream: MediaStream) => {
     if (!this.sfuClient) throw new Error(`Call not joined yet.`);
     // joining is in progress, and we should wait until the client is ready
     await this.sfuClient.joinTask;
@@ -1435,6 +1418,9 @@ export class Call {
     if (!this.trackPublishOrder.includes(TrackType.SCREEN_SHARE)) {
       this.trackPublishOrder.push(TrackType.SCREEN_SHARE);
     }
+    const opts: PublishOptions = {
+      screenShareSettings: this.screenShare.getSettings(),
+    };
     await this.publisher.publishStream(
       screenShareStream,
       screenShareTrack,
@@ -1470,6 +1456,16 @@ export class Call {
     );
     await this.publisher?.unpublishStream(trackType, stopTrack);
   };
+
+  /**
+   * Updates the preferred publishing options
+   *
+   * @internal
+   * @param options the options to use.
+   */
+  updatePublishOptions(options: PublishOptions) {
+    this.publishOptions = { ...this.publishOptions, ...options };
+  }
 
   /**
    * Notifies the SFU that a noise cancellation process has started.
@@ -1531,16 +1527,6 @@ export class Call {
    */
   setSortParticipantsBy: CallState['setSortParticipantsBy'] = (criteria) => {
     return this.state.setSortParticipantsBy(criteria);
-  };
-
-  /**
-   * Updates the list of video layers to publish.
-   *
-   * @internal
-   * @param enabledLayers the list of layers to enable.
-   */
-  updatePublishQuality = async (enabledLayers: VideoLayerSetting[]) => {
-    return this.publisher?.updateVideoPublishQuality(enabledLayers);
   };
 
   /**
@@ -2143,10 +2129,7 @@ export class Call {
         this.camera.state.mediaStream &&
         !this.publisher?.isPublishing(TrackType.VIDEO)
       ) {
-        await this.publishVideoStream(
-          this.camera.state.mediaStream,
-          this.camera.publishOptions,
-        );
+        await this.publishVideoStream(this.camera.state.mediaStream);
       }
 
       // Start camera if backend config specifies, and there is no local setting