+++ packages/client/src/devices/CameraManager.ts
@@ -4,20 +4,14 @@ import { CameraDirection, CameraManagerState } from './CameraManagerState';
 import { InputMediaDeviceManager } from './InputMediaDeviceManager';
 import { getVideoDevices, getVideoStream } from './devices';
 import { TrackType } from '../gen/video/sfu/models/models';
-import { PreferredCodec, PublishOptions } from '../types';
+import { PreferredCodec } from '../types';
+import { isMobile } from '../compatibility';
 
 export class CameraManager extends InputMediaDeviceManager<CameraManagerState> {
   private targetResolution = {
     width: 1280,
     height: 720,
   };
-
-  /**
-   * The options to use when publishing the video stream.
-   *
-   * @internal
-   */
-  publishOptions: PublishOptions | undefined;
 
   /**
    * Constructs a new CameraManager.
@@ -34,10 +28,14 @@ export class CameraManager extends InputMediaDeviceManager<CameraManagerState> {
    * @param direction the direction of the camera to select.
    */
   async selectDirection(direction: Exclude<CameraDirection, undefined>) {
-    this.state.setDirection(direction);
-    // Providing both device id and direction doesn't work, so we deselect the device
-    this.state.setDevice(undefined);
-    await this.applySettingsToStream();
+    if (isMobile()) {
+      this.state.setDirection(direction);
+      // Providing both device id and direction doesn't work, so we deselect the device
+      this.state.setDevice(undefined);
+      await this.applySettingsToStream();
+    } else {
+      this.logger('warn', 'Camera direction ignored for desktop devices');
+    }
   }
 
   /**
@@ -65,10 +63,10 @@ export class CameraManager extends InputMediaDeviceManager<CameraManagerState> {
         this.logger('warn', 'could not apply target resolution', error);
       }
     }
-    if (this.enabled) {
-      const { width, height } = this.state
-        .mediaStream!.getVideoTracks()[0]
-        ?.getSettings();
+    if (this.enabled && this.state.mediaStream) {
+      const [videoTrack] = this.state.mediaStream.getVideoTracks();
+      if (!videoTrack) return;
+      const { width, height } = videoTrack.getSettings();
       if (
         width !== this.targetResolution.width ||
         height !== this.targetResolution.height
@@ -86,38 +84,11 @@ export class CameraManager extends InputMediaDeviceManager<CameraManagerState> {
    * Sets the preferred codec for encoding the video.
    *
    * @internal internal use only, not part of the public API.
+   * @deprecated use {@link call.updatePublishOptions} instead.
    * @param codec the codec to use for encoding the video.
    */
   setPreferredCodec(codec: PreferredCodec | undefined) {
-    this.updatePublishOptions({ preferredCodec: codec });
-  }
-
-  /**
-   * Updates the preferred publish options for the video stream.
-   *
-   * @internal
-   * @param options the options to use.
-   */
-  updatePublishOptions(options: PublishOptions) {
-    this.publishOptions = { ...this.publishOptions, ...options };
-  }
-
-  /**
-   * Returns the capture resolution of the camera.
-   */
-  getCaptureResolution() {
-    const { mediaStream } = this.state;
-    if (!mediaStream) return;
-
-    const [videoTrack] = mediaStream.getVideoTracks();
-    if (!videoTrack) return;
-
-    const settings = videoTrack.getSettings();
-    return {
-      width: settings.width,
-      height: settings.height,
-      frameRate: settings.frameRate,
-    };
+    this.call.updatePublishOptions({ preferredCodec: codec });
   }
 
   protected getDevices(): Observable<MediaDeviceInfo[]> {
@@ -131,7 +102,7 @@ export class CameraManager extends InputMediaDeviceManager<CameraManagerState> {
     constraints.height = this.targetResolution.height;
     // We can't set both device id and facing mode
     // Device id has higher priority
-    if (!constraints.deviceId && this.state.direction) {
+    if (!constraints.deviceId && this.state.direction && isMobile()) {
       constraints.facingMode =
         this.state.direction === 'front' ? 'user' : 'environment';
     }
@@ -139,7 +110,7 @@ export class CameraManager extends InputMediaDeviceManager<CameraManagerState> {
   }
 
   protected publishStream(stream: MediaStream): Promise<void> {
-    return this.call.publishVideoStream(stream, this.publishOptions);
+    return this.call.publishVideoStream(stream);
   }
 
   protected stopPublishStream(stopTracks: boolean): Promise<void> {