+++ packages/client/src/types.ts
@@ -165,12 +165,40 @@ export type SubscriptionChanges = {
  * A preferred codec to use when publishing a video track.
  * @internal
  */
-export type PreferredCodec = 'vp8' | 'h264' | string;
+export type PreferredCodec = 'vp8' | 'h264' | 'vp9' | 'av1';
 
+/**
+ * A collection of track publication options.
+ * @internal
+ */
 export type PublishOptions = {
-  preferredCodec?: PreferredCodec | null;
+  /**
+   * The preferred codec to use when publishing the video stream.
+   */
+  preferredCodec?: PreferredCodec;
+  /**
+   * Force the codec to use when publishing the video stream.
+   * This will override the preferred codec and the internal codec selection logic.
+   * Use with caution.
+   */
+  forceCodec?: PreferredCodec;
+  /**
+   * The preferred scalability to use when publishing the video stream.
+   * Applicable only for SVC codecs.
+   */
+  scalabilityMode?: string;
+  /**
+   * The preferred bitrate to use when publishing the video stream.
+   */
   preferredBitrate?: number;
+  /**
+   * The preferred downscale factor to use when publishing the video stream
+   * in simulcast mode (non-SVC).
+   */
   bitrateDownscaleFactor?: number;
+  /**
+   * Screen share settings.
+   */
   screenShareSettings?: ScreenShareSettings;
 };