+++ sample-apps/react-native/dogfood/src/components/ParticipantsInfoList.tsx
@@ -108,6 +108,7 @@ export const ParticipantsInfoList = ({
       transparent
       visible={isCallParticipantsInfoVisible}
       onRequestClose={onCloseCallParticipantsVisible}
+      supportedOrientations={['portrait', 'landscape']}
     >
       <>
         {/*independent background, needed due to desired opacity only
@@ -154,6 +155,7 @@ export const ParticipantsInfoList = ({
           transparent
           visible={!!selectedParticipant}
           onRequestClose={() => setSelectedParticipant(undefined)}
+          supportedOrientations={['portrait', 'landscape']}
         >
           <>
             {/*independent background, needed due to desired opacity only