+++ sample-apps/react-native/dogfood/src/components/VideoEffectsButton/index.tsx
@@ -56,6 +56,7 @@ const FilterButton = () => {
         transparent
         visible={modalVisible}
         onRequestClose={closeModal}
+        supportedOrientations={['portrait', 'landscape']}
       >
         <Pressable style={styles.centeredView} onPress={closeModal}>
           <View style={styles.modalView} onStartShouldSetResponder={() => true}>